+++ src/protocols/minecraft/protocol/legacy_bv1_8.rs
@@ -51,9 +51,9 @@ impl LegacyBV1_8 {
         Ok(Response {
             version_name: "Beta 1.8+".to_string(),
             version_protocol: -1,
-            max_players,
-            online_players,
-            sample_players: None,
+            players_maximum: max_players,
+            players_online: online_players,
+            players_sample: None,
             description,
             favicon: None,
             previews_chat: None,