+++ src/protocols/minecraft/protocol/legacy_v1_4.rs
@@ -56,9 +56,9 @@ impl LegacyV1_4 {
         Ok(Response {
             version_name: "1.4+".to_string(),
             version_protocol: -1,
-            max_players,
-            online_players,
-            sample_players: None,
+            players_maximum: max_players,
+            players_online: online_players,
+            players_sample: None,
             description,
             favicon: None,
             previews_chat: None,