+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 Who knows what the future holds...
 
+# X - DD/MM/YYYY
+### Changes:
+[Risk of Rain 2](https://store.steampowered.com/app/632360/Risk_of_Rain_2/) support.
+
+### Breaking:
+Nothing (yet).
+
 # 0.0.7 - 03/01/2023
 ### Changes:
 [Minecraft](https://www.minecraft.com) bedrock edition support.