+++ CHANGELOG.md
@@ -5,13 +5,14 @@ Who knows what the future holds...
 ### Changes:
 Games: 
 - [Risk of Rain 2](https://store.steampowered.com/app/632360/Risk_of_Rain_2/) support.
-- [Battalion 1944](https://store.steampowered.com/app/489940/BATTALION_Legacy/) support.
+- [Battalion 1944](https://store.steampowered.com/app/489940/BATTALION_Legacy/) support. 
+- [Black Mesa](https://store.steampowered.com/app/362890/Black_Mesa/) support.
 
 Protocols:
 - Valve: Players with no name are no more added to the `players_details` field.
 
 Crate: 
-- `MSRV` is now `1.56.1` (from `1.58.1`)
+- `MSRV` is now `1.56.1` (was `1.58.1`)
 
 ### Breaking:
 Protocols: